move ringbuffer
